backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
anthropic>=0.20.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
websockets>=12.0
aiofiles>=23.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.35
faker==24.0.0
python-multipart>=0.0.7
